Crate presage

Re-exports

pub use libsignal_service;

Modules

prelude

Structs

Confirmation

ContentProto

Linking

Manager

Registered

Registration

RegistrationOptions

SledStore

Enums

Error

MigrationConflictStrategy

Sometimes Migrations can’t proceed without having to drop existing data. This allows you to configure, how these cases should be handled.

Thread

A thread specifies where a message was sent, either to or from a contact or in a group.

Traits

ContactsStore

MessageStore

A MessageStore can store messages in the form Content and retrieve messages either by [MessageIdentity], by Thread or completely.

StateStore

Store