Module presage::manager

Signal manager and its states

Structs

Confirmation

Manager state after a successful registration of new main device

Linking

Manager state where it is possible to link a new secondary device

Manager

Signal manager

Registered

Manager state when the client is registered and can send and receive messages from Signal

Registration

Manager state where it is possible to register a new main device

RegistrationData

Registration data like device name, and credentials to connect to Signal

RegistrationOptions

Options when registering a new main device

Enums

ReceivingMode

The mode receiving messages stream