Struct presage::prelude::PhoneNumber

pub struct PhoneNumber { /* private fields */ }

A phone number.

Implementations

impl PhoneNumber

pub fn country(&self) -> Country<'_>

Get information about the country for the phone number.

pub fn code(&self) -> &Code

Get the country code.

pub fn national(&self) -> &NationalNumber

Get the national number.

pub fn extension(&self) -> Option<&Extension>

Get the extension.

pub fn carrier(&self) -> Option<&Carrier>

Get the carrier.

pub fn format(&self) -> Formatter<'_, 'static, 'static>

Prepare a formatter for this PhoneNumber.

Example

use phonenumber::{self, country, Mode};

let number = phonenumber::parse(Some(country::DE), "301/23456").unwrap()
    .format().mode(Mode::National).to_string();

assert_eq!("030 123456", number);

pub fn format_with<'n, 'd>( &'n self, database: &'d Database ) -> Formatter<'n, 'd, 'static>

Prepare a formatter for this PhoneNumber with the given Database.

pub fn metadata<'a>(&self, database: &'a Database) -> Option<&'a Metadata>

Get the metadata that applies to this phone number from the given database.

pub fn is_valid(&self) -> bool

Check if the phone number is valid.

pub fn is_valid_with(&self, database: &Database) -> bool

Check if the phone number is valid with the given Database.