presage::proto

Struct AccountRecord

pub struct AccountRecord {

    pub profile_key: Vec<u8>,
    pub given_name: String,
    pub family_name: String,
    pub avatar_url_path: String,
    pub note_to_self_archived: bool,
    pub read_receipts: bool,
    pub sealed_sender_indicators: bool,
    pub typing_indicators: bool,
    pub note_to_self_marked_unread: bool,
    pub link_previews: bool,
    pub phone_number_sharing_mode: i32,
    pub unlisted_phone_number: bool,
    pub pinned_conversations: Vec<PinnedConversation>,
    pub prefer_contact_avatars: bool,
    pub payments: Option<Payments>,
    pub universal_expire_timer: u32,
    pub primary_sends_sms: bool,
    pub preferred_reaction_emoji: Vec<String>,
    pub subscriber_id: Vec<u8>,
    pub subscriber_currency_code: String,
    pub display_badges_on_profile: bool,
    pub subscription_manually_cancelled: bool,
    pub keep_muted_chats_archived: bool,
    pub has_set_my_stories_privacy: bool,
    pub has_viewed_onboarding_story: bool,
    pub stories_disabled: bool,
    pub story_view_receipts_enabled: i32,
    pub has_seen_group_story_education_sheet: bool,
    pub username: String,
    pub has_completed_username_onboarding: bool,
    pub username_link: Option<UsernameLink>,
    pub backup_tier: Option<u64>,
    pub backup_subscriber_data: Option<IapSubscriberData>,
    pub avatar_color: Option<i32>,
    pub backup_tier_history: Option<BackupTierHistory>,
    pub notification_profile_manual_override: Option<NotificationProfileManualOverride>,
    pub notification_profile_sync_disabled: bool,
    pub automatic_key_verification_disabled: bool,
    pub has_seen_admin_delete_education_dialog: bool,
    pub release_notes_chat_archived: Option<bool>,
    pub release_notes_chat_muted_until_timestamp: Option<u64>,
    pub release_notes_chat_blocked: Option<bool>,
    pub release_notes_chat_marked_unread: Option<bool>,

}

Fields
profile_key: Vec<u8>given_name: Stringfamily_name: Stringavatar_url_path: Stringnote_to_self_archived: boolread_receipts: boolsealed_sender_indicators: booltyping_indicators: boolnote_to_self_marked_unread: boollink_previews: boolphone_number_sharing_mode: i32unlisted_phone_number: boolpinned_conversations: Vec<PinnedConversation>prefer_contact_avatars: boolpayments: Option<Payments>universal_expire_timer: u32primary_sends_sms: boolpreferred_reaction_emoji: Vec<String>subscriber_id: Vec<u8>subscriber_currency_code: Stringdisplay_badges_on_profile: boolsubscription_manually_cancelled: boolkeep_muted_chats_archived: boolhas_set_my_stories_privacy: boolhas_viewed_onboarding_story: boolstories_disabled: boolstory_view_receipts_enabled: i32has_seen_group_story_education_sheet: boolusername: Stringhas_completed_username_onboarding: boolusername_link: Option<UsernameLink>backup_tier: Option<u64>

See zkgroup for integer particular values. Unset if backups are not enabled.

backup_subscriber_data: Option<IapSubscriberData>avatar_color: Option<i32>backup_tier_history: Option<BackupTierHistory>notification_profile_manual_override: Option<NotificationProfileManualOverride>notification_profile_sync_disabled: boolautomatic_key_verification_disabled: boolhas_seen_admin_delete_education_dialog: boolrelease_notes_chat_archived: Option<bool>release_notes_chat_muted_until_timestamp: Option<u64>release_notes_chat_blocked: Option<bool>release_notes_chat_marked_unread: Option<bool>
Implementations

impl AccountRecord

pub fn phone_number_sharing_mode(&self) -> PhoneNumberSharingMode

Returns the enum value of phone_number_sharing_mode, or the default if the field is set to an invalid enum value.

pub fn set_phone_number_sharing_mode(&mut self, value: PhoneNumberSharingMode)

Sets phone_number_sharing_mode to the provided enum value.

pub fn story_view_receipts_enabled(&self) -> OptionalBool

Returns the enum value of story_view_receipts_enabled, or the default if the field is set to an invalid enum value.

pub fn set_story_view_receipts_enabled(&mut self, value: OptionalBool)

Sets story_view_receipts_enabled to the provided enum value.

pub fn backup_tier(&self) -> u64

Returns the value of backup_tier, or the default value if backup_tier is unset.

pub fn avatar_color(&self) -> AvatarColor

Returns the enum value of avatar_color, or the default if the field is unset or set to an invalid enum value.

pub fn set_avatar_color(&mut self, value: AvatarColor)

Sets avatar_color to the provided enum value.

pub fn release_notes_chat_archived(&self) -> bool

Returns the value of release_notes_chat_archived, or the default value if release_notes_chat_archived is unset.

pub fn release_notes_chat_muted_until_timestamp(&self) -> u64

Returns the value of release_notes_chat_muted_until_timestamp, or the default value if release_notes_chat_muted_until_timestamp is unset.

pub fn release_notes_chat_blocked(&self) -> bool

Returns the value of release_notes_chat_blocked, or the default value if release_notes_chat_blocked is unset.

pub fn release_notes_chat_marked_unread(&self) -> bool

Returns the value of release_notes_chat_marked_unread, or the default value if release_notes_chat_marked_unread is unset.