Module presage::store

Traits that are used by the manager for storing the data.

Enums

Thread

A thread specifies where a message was sent, either to or from a contact or in a group.

Traits

ContentExt

Extension trait of Content

ContentsStore

Stores messages, contacts, groups and profiles

PreKeyStoreExt

Stores the keys published ahead of time, pre-keys

StateStore

Stores the registered state of the manager

Store

The manager store trait combining all other stores into a single one

StoreError

An error trait implemented by store error types