Trait presage::store::ContentsStore

pub trait ContentsStore: Send + Sync {
    type ContentsStoreError: StoreError;
    type ContactsIter: Iterator<Item = Result<Contact, Self::ContentsStoreError>>;
    type GroupsIter: Iterator<Item = Result<(GroupMasterKeyBytes, Group), Self::ContentsStoreError>>;
    type MessagesIter: Iterator<Item = Result<Content, Self::ContentsStoreError>>;
    type StickerPacksIter: Iterator<Item = Result<StickerPack, Self::ContentsStoreError>>;

    // Required methods
    fn clear_messages(&mut self) -> Result<(), Self::ContentsStoreError>;
    fn clear_thread(
        &mut self,
        thread: &Thread
    ) -> Result<(), Self::ContentsStoreError>;
    fn save_message(
        &self,
        thread: &Thread,
        message: Content
    ) -> Result<(), Self::ContentsStoreError>;
    fn delete_message(
        &mut self,
        thread: &Thread,
        timestamp: u64
    ) -> Result<bool, Self::ContentsStoreError>;
    fn message(
        &self,
        thread: &Thread,
        timestamp: u64
    ) -> Result<Option<Content>, Self::ContentsStoreError>;
    fn messages(
        &self,
        thread: &Thread,
        range: impl RangeBounds<u64>
    ) -> Result<Self::MessagesIter, Self::ContentsStoreError>;
    fn clear_contacts(&mut self) -> Result<(), Self::ContentsStoreError>;
    fn save_contact(
        &mut self,
        contacts: &Contact
    ) -> Result<(), Self::ContentsStoreError>;
    fn contacts(&self) -> Result<Self::ContactsIter, Self::ContentsStoreError>;
    fn contact_by_id(
        &self,
        id: &Uuid
    ) -> Result<Option<Contact>, Self::ContentsStoreError>;
    fn clear_groups(&mut self) -> Result<(), Self::ContentsStoreError>;
    fn save_group(
        &self,
        master_key: GroupMasterKeyBytes,
        group: &Group
    ) -> Result<(), Self::ContentsStoreError>;
    fn groups(&self) -> Result<Self::GroupsIter, Self::ContentsStoreError>;
    fn group(
        &self,
        master_key: GroupMasterKeyBytes
    ) -> Result<Option<Group>, Self::ContentsStoreError>;
    fn save_group_avatar(
        &self,
        master_key: GroupMasterKeyBytes,
        avatar: &AvatarBytes
    ) -> Result<(), Self::ContentsStoreError>;
    fn group_avatar(
        &self,
        master_key: GroupMasterKeyBytes
    ) -> Result<Option<AvatarBytes>, Self::ContentsStoreError>;
    fn upsert_profile_key(
        &mut self,
        uuid: &Uuid,
        key: ProfileKey
    ) -> Result<bool, Self::ContentsStoreError>;
    fn profile_key(
        &self,
        uuid: &Uuid
    ) -> Result<Option<ProfileKey>, Self::ContentsStoreError>;
    fn save_profile(
        &mut self,
        uuid: Uuid,
        key: ProfileKey,
        profile: Profile
    ) -> Result<(), Self::ContentsStoreError>;
    fn profile(
        &self,
        uuid: Uuid,
        key: ProfileKey
    ) -> Result<Option<Profile>, Self::ContentsStoreError>;
    fn save_profile_avatar(
        &mut self,
        uuid: Uuid,
        key: ProfileKey,
        profile: &AvatarBytes
    ) -> Result<(), Self::ContentsStoreError>;
    fn profile_avatar(
        &self,
        uuid: Uuid,
        key: ProfileKey
    ) -> Result<Option<AvatarBytes>, Self::ContentsStoreError>;
    fn add_sticker_pack(
        &mut self,
        pack: &StickerPack
    ) -> Result<(), Self::ContentsStoreError>;
    fn sticker_pack(
        &self,
        id: &[u8]
    ) -> Result<Option<StickerPack>, Self::ContentsStoreError>;
    fn remove_sticker_pack(
        &mut self,
        id: &[u8]
    ) -> Result<bool, Self::ContentsStoreError>;
    fn sticker_packs(
        &self
    ) -> Result<Self::StickerPacksIter, Self::ContentsStoreError>;

    // Provided methods
    fn save_trusted_identity_message(
        &self,
        protocol_address: &ProtocolAddress,
        right_identity_key: IdentityKey,
        verified_state: State
    ) { ... }
    fn expire_timer(
        &self,
        thread: &Thread
    ) -> Result<Option<u32>, Self::ContentsStoreError> { ... }
    fn update_expire_timer(
        &mut self,
        thread: &Thread,
        timer: u32
    ) -> Result<(), Self::ContentsStoreError> { ... }

}

Stores messages, contacts, groups and profiles

Required Associated Types

type ContentsStoreError: StoreError

type ContactsIter: Iterator<Item = Result<Contact, Self::ContentsStoreError>>

Iterator over the contacts

type GroupsIter: Iterator<Item = Result<(GroupMasterKeyBytes, Group), Self::ContentsStoreError>>

Iterator over all stored groups

Each items is a tuple consisting of the group master key and its corresponding data.

type MessagesIter: Iterator<Item = Result<Content, Self::ContentsStoreError>>

Iterator over all stored messages

type StickerPacksIter: Iterator<Item = Result<StickerPack, Self::ContentsStoreError>>

Iterator over all stored sticker packs

Required Methods

fn clear_messages(&mut self) -> Result<(), Self::ContentsStoreError>

Clear all stored messages.

fn clear_thread( &mut self, thread: &Thread ) -> Result<(), Self::ContentsStoreError>

Clear the messages in a thread.

fn save_message( &self, thread: &Thread, message: Content ) -> Result<(), Self::ContentsStoreError>

Save a message in a Thread identified by a timestamp.

fn delete_message( &mut self, thread: &Thread, timestamp: u64 ) -> Result<bool, Self::ContentsStoreError>

Delete a single message, identified by its received timestamp from a thread. Useful when you want to delete a message locally only.

fn message( &self, thread: &Thread, timestamp: u64 ) -> Result<Option<Content>, Self::ContentsStoreError>

Retrieve a message from a Thread by its timestamp.

fn messages( &self, thread: &Thread, range: impl RangeBounds<u64> ) -> Result<Self::MessagesIter, Self::ContentsStoreError>

Retrieve all messages from a Thread within a range in time

fn clear_contacts(&mut self) -> Result<(), Self::ContentsStoreError>

Clear all saved synchronized contact data

fn save_contact( &mut self, contacts: &Contact ) -> Result<(), Self::ContentsStoreError>

Save a contact

fn contacts(&self) -> Result<Self::ContactsIter, Self::ContentsStoreError>

Get an iterator on all stored (synchronized) contacts

fn contact_by_id( &self, id: &Uuid ) -> Result<Option<Contact>, Self::ContentsStoreError>

Get contact data for a single user by its Uuid.

fn clear_groups(&mut self) -> Result<(), Self::ContentsStoreError>

Delete all cached group data

fn save_group( &self, master_key: GroupMasterKeyBytes, group: &Group ) -> Result<(), Self::ContentsStoreError>

Save a group in the cache

fn groups(&self) -> Result<Self::GroupsIter, Self::ContentsStoreError>

Get an iterator on all cached groups

fn group( &self, master_key: GroupMasterKeyBytes ) -> Result<Option<Group>, Self::ContentsStoreError>

Retrieve a single unencrypted group indexed by its [GroupMasterKeyBytes]

fn save_group_avatar( &self, master_key: GroupMasterKeyBytes, avatar: &AvatarBytes ) -> Result<(), Self::ContentsStoreError>

Save a group avatar in the cache

fn group_avatar( &self, master_key: GroupMasterKeyBytes ) -> Result<Option<AvatarBytes>, Self::ContentsStoreError>

Retrieve a group avatar from the cache.

fn upsert_profile_key( &mut self, uuid: &Uuid, key: ProfileKey ) -> Result<bool, Self::ContentsStoreError>

Insert or update the profile key of a contact

fn profile_key( &self, uuid: &Uuid ) -> Result<Option<ProfileKey>, Self::ContentsStoreError>

Get the profile key for a contact

fn save_profile( &mut self, uuid: Uuid, key: ProfileKey, profile: Profile ) -> Result<(), Self::ContentsStoreError>

Save a profile by Uuid and ProfileKey.

fn profile( &self, uuid: Uuid, key: ProfileKey ) -> Result<Option<Profile>, Self::ContentsStoreError>

Retrieve a profile by Uuid and ProfileKey.

fn save_profile_avatar( &mut self, uuid: Uuid, key: ProfileKey, profile: &AvatarBytes ) -> Result<(), Self::ContentsStoreError>

Save a profile avatar by Uuid and ProfileKey.

fn profile_avatar( &self, uuid: Uuid, key: ProfileKey ) -> Result<Option<AvatarBytes>, Self::ContentsStoreError>

Retrieve a profile avatar by Uuid and ProfileKey.

fn add_sticker_pack( &mut self, pack: &StickerPack ) -> Result<(), Self::ContentsStoreError>

Stickers Add a sticker pack

fn sticker_pack( &self, id: &[u8] ) -> Result<Option<StickerPack>, Self::ContentsStoreError>

Gets a cached sticker pack

fn remove_sticker_pack( &mut self, id: &[u8] ) -> Result<bool, Self::ContentsStoreError>

Removes a sticker pack

fn sticker_packs( &self ) -> Result<Self::StickerPacksIter, Self::ContentsStoreError>

Get an iterator on all installed stickerpacks

Provided Methods

fn save_trusted_identity_message( &self, protocol_address: &ProtocolAddress, right_identity_key: IdentityKey, verified_state: State )

Saves a message that can show users when the identity of a contact has changed On Signal Android, this is usually displayed as: “Your safety number with XYZ has changed.”

fn expire_timer( &self, thread: &Thread ) -> Result<Option<u32>, Self::ContentsStoreError>

Get the expire timer from a Thread, which corresponds to either Contact::expire_timer or Group::disappearing_messages_timer.

fn update_expire_timer( &mut self, thread: &Thread, timer: u32 ) -> Result<(), Self::ContentsStoreError>

Update the expire timer from a Thread, which corresponds to either Contact::expire_timer or Group::disappearing_messages_timer.

Object Safety

This trait is not object safe.