Struct presage::Manager

pub struct Manager<Store, State> { /* private fields */ }

Implementations

impl<C: Store> Manager<C, Registration>

pub async fn register( config_store: C, registration_options: RegistrationOptions<'_> ) -> Result<Manager<C, Confirmation>, Error<C::Error>>

Registers a new account with a phone number (and some options).

The returned value is a confirmation manager which you then have to use to send the confirmation code.

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    use std::str::FromStr;

    use presage::{
        prelude::{phonenumber::PhoneNumber, SignalServers},
        Manager, MigrationConflictStrategy, RegistrationOptions, SledStore,
    };

    let config_store =
        SledStore::open("/tmp/presage-example", MigrationConflictStrategy::Drop)?;

    let manager = Manager::register(
        config_store,
        RegistrationOptions {
            signal_servers: SignalServers::Production,
            phone_number: PhoneNumber::from_str("+16137827274")?,
            use_voice_call: false,
            captcha: None,
            force: false,
        },
    )
    .await?;

    Ok(())
}

impl<C: Store> Manager<C, Linking>

pub async fn link_secondary_device( config_store: C, signal_servers: SignalServers, device_name: String, provisioning_link_channel: Sender<Url> ) -> Result<Manager<C, Registered>, Error<C::Error>>

Links this client as a secondary device from the device used to register the account (usually a phone). The URL to present to the user will be sent in the channel given as the argument.

use futures::{channel::oneshot, future, StreamExt};
use presage::{prelude::SignalServers, Manager, MigrationConflictStrategy, SledStore};

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    let config_store =
        SledStore::open("/tmp/presage-example", MigrationConflictStrategy::Drop)?;

    let (mut tx, mut rx) = oneshot::channel();
    let (manager, err) = future::join(
        Manager::link_secondary_device(
            config_store,
            SignalServers::Production,
            "my-linked-client".into(),
            tx,
        ),
        async move {
            match rx.await {
                Ok(url) => println!("Show URL {} as QR code to user", url),
                Err(e) => println!("Error linking device: {}", e),
            }
        },
    )
    .await;

    Ok(())
}

impl<C: Store> Manager<C, Confirmation>

pub async fn confirm_verification_code( self, confirm_code: impl AsRef<str> ) -> Result<Manager<C, Registered>, Error<C::Error>>

Confirm a newly registered account using the code you received by SMS or phone call.

Returns a registered manager that you can use to send and receive messages.

impl<C: Store> Manager<C, Registered>

pub async fn load_registered(config_store: C) -> Result<Self, Error<C::Error>>

Loads a previously registered account from the implemented Store.

Returns a instance of Manager you can use to send & receive messages.

pub async fn request_contacts_sync(&mut self) -> Result<(), Error<C::Error>>

Request that the primary device to encrypt & send all of its contacts as a message to ourselves which can be then received, decrypted and stored in the message receiving loop.

Note: If successful, the contacts are not yet received and stored, but will only be processed when they’re received using the MessageReceiver.

pub async fn submit_recaptcha_challenge( &self, token: &str, captcha: &str ) -> Result<(), Error<C::Error>>

pub fn state(&self) -> &Registered

Returns a handle on the registered state

pub async fn whoami(&self) -> Result<WhoAmIResponse, Error<C::Error>>

Fetches basic information on the registered device.

pub async fn retrieve_profile(&mut self) -> Result<Profile, Error<C::Error>>

Fetches the profile (name, about, status emoji) of the registered user.

pub async fn retrieve_profile_by_uuid( &mut self, uuid: Uuid, profile_key: ProfileKey ) -> Result<Profile, Error<C::Error>>

Fetches the profile of the provided user by UUID and profile key.

pub fn contact_by_id( &self, id: &Uuid ) -> Result<Option<Contact>, Error<C::Error>>

Get a single contact by its UUID

Note: this only currently works when linked as secondary device (the contacts are sent by the primary device at linking time)

pub fn contacts( &self ) -> Result<impl Iterator<Item = Result<Contact, Error<C::Error>>>, Error<C::Error>>

Returns an iterator on contacts stored in the Store.

pub fn group( &self, master_key_bytes: &[u8] ) -> Result<Option<Group>, Error<C::Error>>

Get a group (either from the local cache, or fetch it remotely) using its master key

pub fn groups(&self) -> Result<C::GroupsIter, Error<C::Error>>

Returns an iterator on groups stored in the Store.

pub fn message( &self, thread: &Thread, timestamp: u64 ) -> Result<Option<Content>, Error<C::Error>>

Get a single message in a thread (identified by its server-side sent timestamp)

pub fn messages( &self, thread: &Thread, range: impl RangeBounds<u64> ) -> Result<C::MessagesIter, Error<C::Error>>

Get an iterator of messages in a thread, optionally starting from a point in time.

pub async fn receive_messages( &mut self ) -> Result<impl Stream<Item = Content>, Error<C::Error>>

Starts receiving and storing messages.

Returns a [Stream] of messages to consume. Messages will also be stored by the implementation of the [MessageStore].

pub async fn send_message( &mut self, recipient_addr: impl Into<ServiceAddress>, message: impl Into<ContentBody>, timestamp: u64 ) -> Result<(), Error<C::Error>>

Sends a messages to the provided ServiceAddress. The timestamp should be set to now and is used by Signal mobile apps to order messages later, and apply reactions.

This method will automatically update the [DataMessage::expiration_timer] if it is set to None such that the chat will keep the current expire timer.

pub async fn upload_attachments( &self, attachments: Vec<(AttachmentSpec, Vec<u8>)> ) -> Result<Vec<Result<AttachmentPointer, AttachmentUploadError>>, Error<C::Error>>

Uploads attachments prior to linking them in a message.

pub async fn send_message_to_group( &mut self, master_key_bytes: &[u8], message: DataMessage, timestamp: u64 ) -> Result<(), Error<C::Error>>

Sends one message in a group (v2). The master_key_bytes is required to have 32 elements.

This method will automatically update the [DataMessage::expiration_timer] if it is set to None such that the chat will keep the current expire timer.

pub async fn clear_sessions( &self, recipient: &ServiceAddress ) -> Result<(), Error<C::Error>>

Clears all sessions established wiht recipient.

pub async fn get_attachment( &self, attachment_pointer: &AttachmentPointer ) -> Result<Vec<u8>, Error<C::Error>>

Downloads and decrypts a single attachment.

pub async fn send_session_reset( &mut self, recipient: &ServiceAddress, timestamp: u64 ) -> Result<(), Error<C::Error>>

pub fn get_contacts( &self ) -> Result<impl Iterator<Item = Result<Contact, Error<C::Error>>>, Error<C::Error>>

👎Deprecated: use Manager::contact_by_id

pub fn get_contact_by_id( &self, id: Uuid ) -> Result<Option<Contact>, Error<C::Error>>

👎Deprecated: use Manager::contact_by_id

pub fn get_groups(&self) -> Result<C::GroupsIter, Error<C::Error>>

👎Deprecated: use Manager::groups

pub fn get_group( &self, master_key_bytes: &[u8] ) -> Result<Option<Group>, Error<C::Error>>

👎Deprecated: use Manager::group