Trait presage::MessageStore

pub trait MessageStore {
    type MessagesIter: Iterator<Item = Content>;

    fn save_message(
        &mut self, 
        message: Content, 
        receiver: Option<impl Into<ServiceAddress>>
    ) -> Result<(), Error>;
    fn message(
        &self, 
        thread: &Thread, 
        timestamp: u64
    ) -> Result<Option<Content>, Error>;
    fn messages(
        &self, 
        thread: &Thread, 
        from: Option<u64>
    ) -> Result<Self::MessagesIter, Error>;
}

A MessageStore can store messages in the form Content and retrieve messages either by [MessageIdentity], by Thread or completly.

Required Associated Types

type MessagesIter: Iterator<Item = Content>

Required Methods

fn save_message(
&mut self,
message: Content,
receiver: Option<impl Into<ServiceAddress>>
) -> Result<(), Error>

Save a message. The receiver-argument specifies the ServiceAddress of the receiver of that message and is needed for correct association of the message to a Thread. If that message was received, it should be None.

fn message(
&self,
thread: &Thread,
timestamp: u64
) -> Result<Option<Content>, Error>

Retrieve a message by its a Thread and its timestamp.

fn messages(
&self,
thread: &Thread,
from: Option<u64>
) -> Result<Self::MessagesIter, Error>

Retrieve a message by a Thread.