Trait presage::Store

pub trait Store: PreKeyStore + SignedPreKeyStore + SessionStoreExt + IdentityKeyStore + StateStore<Registered> + ContactsStore + MessageStore + SenderKeyStore + Sync + Clone {
    fn clear(&mut self) -> Result<(), Error>;
    fn pre_keys_offset_id(&self) -> Result<u32, Error>;
    fn set_pre_keys_offset_id(&mut self, id: u32) -> Result<(), Error>;
    fn next_signed_pre_key_id(&self) -> Result<u32, Error>;
    fn set_next_signed_pre_key_id(&mut self, id: u32) -> Result<(), Error>;
}

Required Methods

fn clear(&mut self) -> Result<(), Error>

Clear the entire store, this can be useful when re-initializing an existing client Note: you can implement this the way you want and only clear the database partially but should always make sure the state and all keys are gone.

fn pre_keys_offset_id(&self) -> Result<u32, Error>

fn set_pre_keys_offset_id(&mut self, id: u32) -> Result<(), Error>

fn next_signed_pre_key_id(&self) -> Result<u32, Error>

fn set_next_signed_pre_key_id(&mut self, id: u32) -> Result<(), Error>