Crate presage

Re-exports

pub use libsignal_service;

Modules

prelude

Structs

Confirmation

ContentProto

Linking

Manager

Registered

Registration

RegistrationOptions

SecretVolatileConfigStore

SledConfigStore

VolatileConfigStore

Enums

Error

Thread

A thread specifies where a message was sent, either to or from a contact or in a group.

Traits

ConfigStore

ContactsStore

MessageStore

A MessageStore can store messages in the form Content and retrieve messages either by [MessageIdentity], by Thread or completly.

StateStore