Struct presage::manager::Manager

pub struct Manager<Store, State> { /* private fields */ }

Signal manager

The manager is parametrized over the crate::store::Store which stores the configuration, keys and optionally messages; and over the type state which describes what is the current state of the manager: in registration, linking, TODO

Depending on the state specific methods are available or not.

Implementations

impl<S: Store> Manager<S, Confirmation>

pub async fn confirm_verification_code( self, confirmation_code: impl AsRef<str> ) -> Result<Manager<S, Registered>, Error<S::Error>>

Confirm a newly registered account using the code you received by SMS or phone call.

Returns a registered manager that you can use to send and receive messages.

impl<S: Store> Manager<S, Linking>

pub async fn link_secondary_device( store: S, signal_servers: SignalServers, device_name: String, provisioning_link_channel: Sender<Url> ) -> Result<Manager<S, Registered>, Error<S::Error>>

Links this client as a secondary device from the device used to register the account (usually a phone). The URL to present to the user will be sent in the channel given as the argument.

use futures::{channel::oneshot, future, StreamExt};
use presage::libsignal_service::configuration::SignalServers;
use presage::Manager;
use presage_store_sled::{MigrationConflictStrategy, OnNewIdentity, SledStore};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    let store =
        SledStore::open("/tmp/presage-example", MigrationConflictStrategy::Drop, OnNewIdentity::Trust)?;

    let (mut tx, mut rx) = oneshot::channel();
    let (manager, err) = future::join(
        Manager::link_secondary_device(
            store,
            SignalServers::Production,
            "my-linked-client".into(),
            tx,
        ),
        async move {
            match rx.await {
                Ok(url) => println!("Show URL {} as QR code to user", url),
                Err(e) => println!("Error linking device: {}", e),
            }
        },
    )
    .await;

    Ok(())
}

impl<S: Store> Manager<S, Registered>

pub async fn load_registered(store: S) -> Result<Self, Error<S::Error>>

Loads a previously registered account from the implemented Store.

Returns a instance of Manager you can use to send & receive messages.

pub fn store(&self) -> &S

Returns a handle to the Store implementation.

pub fn registration_data(&self) -> &RegistrationData

Returns a handle on the RegistrationData.

pub async fn sync_contacts(&mut self) -> Result<(), Error<S::Error>>

Requests contacts synchronization and waits until the primary device sends them

Note: DO NOT call this function if you’re already running a receiving loop

pub async fn request_contacts(&mut self) -> Result<(), Error<S::Error>>

Request the primary device to encrypt & send all of its contacts.

Note: If successful, the contacts are not yet received and stored, but will only be processed when they’re received after polling on the

pub async fn submit_recaptcha_challenge( &self, token: &str, captcha: &str ) -> Result<(), Error<S::Error>>

pub async fn whoami(&self) -> Result<WhoAmIResponse, Error<S::Error>>

Fetches basic information on the registered device.

pub async fn retrieve_profile(&mut self) -> Result<Profile, Error<S::Error>>

Fetches the profile (name, about, status emoji) of the registered user.

pub async fn retrieve_profile_by_uuid( &mut self, uuid: Uuid, profile_key: ProfileKey ) -> Result<Profile, Error<S::Error>>

Fetches the profile of the provided user by UUID and profile key.

pub fn messages( &self, thread: &Thread, range: impl RangeBounds<u64> ) -> Result<S::MessagesIter, Error<S::Error>>

Get an iterator of messages in a thread, optionally starting from a point in time.

pub async fn receive_messages( &mut self, mode: ReceivingMode ) -> Result<impl Stream<Item = Content>, Error<S::Error>>

Starts receiving and storing messages.

As a client, it is heavily recommended to run this once in ReceivingMode::InitialSync once before enabling the possiblity of sending messages. That way, all possible updates (sessions, profile keys, sender keys) are processed before trying to encrypt and send messages which might fail otherwise.

Returns a [futures::Stream] of messages to consume. Messages will also be stored by the implementation of the Store.

pub async fn send_message( &mut self, recipient_addr: impl Into<ServiceAddress>, message: impl Into<ContentBody>, timestamp: u64 ) -> Result<(), Error<S::Error>>

Sends a messages to the provided ServiceAddress. The timestamp should be set to now and is used by Signal mobile apps to order messages later, and apply reactions.

This method will automatically update the DataMessage::expire_timer if it is set to None such that the chat will keep the current expire timer.

pub async fn upload_attachments( &self, attachments: Vec<(AttachmentSpec, Vec<u8>)> ) -> Result<Vec<Result<AttachmentPointer, AttachmentUploadError>>, Error<S::Error>>

Uploads attachments prior to linking them in a message.

pub async fn send_message_to_group( &mut self, master_key_bytes: &[u8], message: impl Into<ContentBody>, timestamp: u64 ) -> Result<(), Error<S::Error>>

Sends one message in a group (v2). The master_key_bytes is required to have 32 elements.

This method will automatically update the DataMessage::expire_timer if it is set to None such that the chat will keep the current expire timer.

pub async fn clear_sessions( &self, recipient: &ServiceAddress ) -> Result<(), Error<S::Error>>

Clears all sessions established wiht recipient.

pub async fn get_attachment( &self, attachment_pointer: &AttachmentPointer ) -> Result<Vec<u8>, Error<S::Error>>

Downloads and decrypts a single attachment.

pub async fn send_session_reset( &mut self, recipient: &ServiceAddress, timestamp: u64 ) -> Result<(), Error<S::Error>>

pub async fn thread_title( &self, thread: &Thread ) -> Result<String, Error<S::Error>>

Returns the title of a thread (contact or group).

pub fn contact_by_id( &self, id: &Uuid ) -> Result<Option<Contact>, Error<S::Error>>

👎Deprecated: use the store handle directly

Deprecated methods Get a single contact by its UUID

Note: this only currently works when linked as secondary device (the contacts are sent by the primary device at linking time)

pub fn contacts( &self ) -> Result<impl Iterator<Item = Result<Contact, Error<S::Error>>>, Error<S::Error>>

👎Deprecated: use the store handle directly

Returns an iterator on contacts stored in the Store.

pub fn group( &self, master_key_bytes: &[u8] ) -> Result<Option<Group>, Error<S::Error>>

👎Deprecated: use the store handle directly

Get a group (either from the local cache, or fetch it remotely) using its master key

pub fn groups(&self) -> Result<S::GroupsIter, Error<S::Error>>

👎Deprecated: use the store handle directly

Returns an iterator on groups stored in the Store.

pub fn message( &self, thread: &Thread, timestamp: u64 ) -> Result<Option<Content>, Error<S::Error>>

👎Deprecated: use the store handle directly

Get a single message in a thread (identified by its server-side sent timestamp)

impl<S: Store> Manager<S, Registration>

pub async fn register( store: S, registration_options: RegistrationOptions<'_> ) -> Result<Manager<S, Confirmation>, Error<S::Error>>

Registers a new account with a phone number (and some options).

The returned value is a confirmation manager which you then have to use to send the confirmation code.

use std::str::FromStr;

use presage::libsignal_service::{
    configuration::SignalServers, prelude::phonenumber::PhoneNumber,
};
use presage::manager::RegistrationOptions;
use presage::Manager;
use presage_store_sled::{MigrationConflictStrategy, OnNewIdentity, SledStore};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    let store =
        SledStore::open("/tmp/presage-example", MigrationConflictStrategy::Drop, OnNewIdentity::Trust)?;

    let manager = Manager::register(
        store,
        RegistrationOptions {
            signal_servers: SignalServers::Production,
            phone_number: PhoneNumber::from_str("+16137827274")?,
            use_voice_call: false,
            captcha: None,
            force: false,
        },
    )
    .await?;

    Ok(())
}