presage::manager

Struct Manager

pub struct Manager<Store, State> { /* private fields */ }

Signal manager

The manager is parametrized over the crate::store::Store which stores the configuration, keys and optionally messages; and over the type state which describes what is the current state of the manager: in registration, linking, TODO

Depending on the state specific methods are available or not.

Implementations

impl<S: Store> Manager<S, Confirmation>

pub async fn confirm_verification_code( self, confirmation_code: impl AsRef<str>, ) -> Result<Manager<S, Registered>, Error<S::Error>>

Confirm a newly registered account using the code you received by SMS or phone call.

Returns a registered manager that you can use to send and receive messages.

impl<S: Store> Manager<S, Linking>

pub async fn link_secondary_device( store: S, signal_servers: SignalServers, device_name: String, provisioning_link_channel: Sender<Url>, ) -> Result<Manager<S, Registered>, Error<S::Error>>

Links this client as a secondary device from the device used to register the account (usually a phone). The URL to present to the user will be sent in the channel given as the argument.

use futures::{channel::oneshot, future, StreamExt};
use presage::libsignal_service::configuration::SignalServers;
use presage::Manager;
use presage::model::identity::OnNewIdentity;
use presage_store_sled::{MigrationConflictStrategy, SledStore};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    let store =
        SledStore::open("/tmp/presage-example", MigrationConflictStrategy::Drop, OnNewIdentity::Trust).await?;

    let (mut tx, mut rx) = oneshot::channel();
    let (manager, err) = future::join(
        Manager::link_secondary_device(
            store,
            SignalServers::Production,
            "my-linked-client".into(),
            tx,
        ),
        async move {
            match rx.await {
                Ok(url) => println!("Show URL {} as QR code to user", url),
                Err(e) => println!("Error linking device: {}", e),
            }
        },
    )
    .await;

    Ok(())
}

impl<S: Store> Manager<S, Registered>

pub async fn load_registered(store: S) -> Result<Self, Error<S::Error>>

Loads a previously registered account from the implemented Store.

Returns a instance of Manager you can use to send & receive messages.

pub fn store(&self) -> &S

Returns a handle to the Store implementation.

pub fn registration_data(&self) -> &RegistrationData

Returns a handle on the RegistrationData.

pub async fn request_contacts(&mut self) -> Result<(), Error<S::Error>>

Request the primary device to encrypt & send all of its contacts.

Note: If successful, the contacts are not yet received and stored, but will only be processed when they’re received after polling on the

pub async fn submit_recaptcha_challenge( &self, token: &str, captcha: &str, ) -> Result<(), Error<S::Error>>

pub async fn whoami(&self) -> Result<WhoAmIResponse, Error<S::Error>>

Fetches basic information on the registered device.

pub fn device_id(&self) -> u32

pub async fn retrieve_profile(&mut self) -> Result<Profile, Error<S::Error>>

Fetches the profile (name, about, status emoji) of the registered user.

pub async fn retrieve_profile_by_uuid( &mut self, aci: impl Into<Aci>, profile_key: ProfileKey, ) -> Result<Profile, Error<S::Error>>

Fetches the profile of the provided user by UUID and profile key.

pub async fn retrieve_group_avatar( &mut self, context: GroupContextV2, ) -> Result<Option<AvatarBytes>, Error<S::Error>>

pub async fn retrieve_profile_avatar_by_uuid( &mut self, uuid: Uuid, profile_key: ProfileKey, ) -> Result<Option<AvatarBytes>, Error<S::Error>>

pub async fn receive_messages( &mut self, ) -> Result<impl Stream<Item = Received>, Error<S::Error>>

Starts receiving and storing messages.

As a client, it is heavily recommended to process incoming messages and wait for the Received::QueueEmpty messages until giving the ability for users to send messages. That way, all possible updates (sessions, profile keys, sender keys) are processed before trying to encrypt and send messages, which might get rejected by recipients otherwise.

Returns a [futures::Stream] of messages to consume. Messages will also be stored by the implementation of the Store.

pub async fn send_message( &mut self, recipient: impl Into<ServiceId>, message: impl Into<ContentBody>, timestamp: u64, ) -> Result<(), Error<S::Error>>

Sends a messages to the provided [ServiceId]. The timestamp should be set to now and is used by Signal mobile apps to order messages later, and apply reactions.

This method will automatically update the DataMessage::expire_timer if it is set to None such that the chat will keep the current expire timer. If the expire timer is set, it will be used as is, and the expire timer version will be incremented.

pub async fn upload_attachments( &self, attachments: Vec<(AttachmentSpec, Vec<u8>)>, ) -> Result<Vec<Result<AttachmentPointer, AttachmentUploadError>>, Error<S::Error>>

Uploads attachments prior to linking them in a message.

pub async fn send_message_to_group( &mut self, master_key_bytes: &[u8], message: impl Into<ContentBody>, timestamp: u64, ) -> Result<(), Error<S::Error>>

Sends one message in a group (v2). The master_key_bytes is required to have 32 elements.

This method will automatically update the DataMessage::expire_timer if it is set to None such that the chat will keep the current expire timer.

pub async fn clear_sessions( &self, recipient: &ServiceId, ) -> Result<(), Error<S::Error>>

Clears all sessions established with recipient.

pub async fn get_attachment( &self, attachment_pointer: &AttachmentPointer, ) -> Result<Vec<u8>, Error<S::Error>>

Downloads and decrypts a single attachment.

pub async fn sticker_metadata( &mut self, pack_id: &[u8], sticker_id: u32, ) -> Result<Option<Sticker>, Error<S::Error>>

Gets the metadata of a sticker

pub async fn install_sticker_pack( &mut self, pack_id: &[u8], pack_key: &[u8], ) -> Result<(), Error<S::Error>>

Installs a sticker pack and notifies other registered devices

pub async fn remove_sticker_pack( &mut self, pack_id: &[u8], pack_key: &[u8], ) -> Result<(), Error<S::Error>>

Removes an installed sticker pack

pub async fn send_session_reset( &mut self, recipient: &ServiceId, timestamp: u64, ) -> Result<(), Error<S::Error>>

pub async fn thread_title( &self, thread: &Thread, ) -> Result<String, Error<S::Error>>

Returns the title of a thread (contact or group).

pub fn registration_type(&self) -> RegistrationType

Returns how this client was registered, either as a primary or secondary device.

pub async fn link_secondary( &mut self, secondary: Url, ) -> Result<(), Error<S::Error>>

As a primary device, link a secondary device.

pub async fn unlink_secondary( &self, device_id: i64, ) -> Result<(), Error<S::Error>>

As a primary device, unlink a secondary device.

pub async fn devices(&self) -> Result<Vec<DeviceInfo>, Error<S::Error>>

As a primary device, list all the devices (including the current device).

impl<S: Store> Manager<S, Registration>

pub async fn register( store: S, registration_options: RegistrationOptions<'_>, ) -> Result<Manager<S, Confirmation>, Error<S::Error>>

Registers a new account with a phone number (and some options).

The returned value is a confirmation manager which you then have to use to send the confirmation code.

use std::str::FromStr;

use presage::libsignal_service::{
    configuration::SignalServers, prelude::phonenumber::PhoneNumber,
};
use presage::manager::RegistrationOptions;
use presage::Manager;
use presage::model::identity::OnNewIdentity;

use presage_store_sled::{MigrationConflictStrategy, SledStore};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    let store =
        SledStore::open("/tmp/presage-example", MigrationConflictStrategy::Drop, OnNewIdentity::Trust).await?;

    let manager = Manager::register(
        store,
        RegistrationOptions {
            signal_servers: SignalServers::Production,
            phone_number: PhoneNumber::from_str("+16137827274")?,
            use_voice_call: false,
            captcha: None,
            force: false,
        },
    )
    .await?;

    Ok(())
}