presage

Module store

Traits that are used by the manager for storing the data.

Structs

Sticker

equivalent to Sticker

StickerPack

StickerPackManifest

equivalent to Pack

Enums

Thread

A thread specifies where a message was sent, either to or from a contact or in a group.

Traits

ContentExt

Extension trait of Content

ContentsStore

Stores messages, contacts, groups and profiles

StateStore

Stores the registered state of the manager

Store

The manager store trait combining all other stores into a single one

StoreError

An error trait implemented by store error types

Functions

save_trusted_identity_message

Saves a message that can show users when the identity of a contact has changed On Signal Android, this is usually displayed as: “Your safety number with XYZ has changed.”