presage::store

Function save_trusted_identity_message

pub async fn save_trusted_identity_message<S: Store>(
    store: &S,
    protocol_address: &ProtocolAddress,
    right_identity_key: IdentityKey,
    verified_state: State,
) -> Result<(), S::Error>

Saves a message that can show users when the identity of a contact has changed On Signal Android, this is usually displayed as: “Your safety number with XYZ has changed.”