presage::store

Trait ContentsStore

pub trait ContentsStore: Send + Sync {
    type ContentsStoreError: StoreError;
    type ContactsIter: Iterator<Item = Result<Contact, Self::ContentsStoreError>>;
    type GroupsIter: Iterator<Item = Result<(GroupMasterKeyBytes, Group), Self::ContentsStoreError>>;
    type MessagesIter: Iterator<Item = Result<Content, Self::ContentsStoreError>>;
    type StickerPacksIter: Iterator<Item = Result<StickerPack, Self::ContentsStoreError>>;

    // Required methods
    fn clear_profiles(
        &mut self,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn clear_contents(
        &mut self,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn clear_messages(
        &mut self,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn clear_thread(
        &mut self,
        thread: &Thread,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn save_message(
        &self,
        thread: &Thread,
        message: Content,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn delete_message(
        &mut self,
        thread: &Thread,
        timestamp: u64,
    ) -> impl Future<Output = Result<bool, Self::ContentsStoreError>>;
    fn message(
        &self,
        thread: &Thread,
        timestamp: u64,
    ) -> impl Future<Output = Result<Option<Content>, Self::ContentsStoreError>>;
    fn messages(
        &self,
        thread: &Thread,
        range: impl RangeBounds<u64>,
    ) -> impl Future<Output = Result<Self::MessagesIter, Self::ContentsStoreError>>;
    fn clear_contacts(
        &mut self,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn save_contact(
        &mut self,
        contacts: &Contact,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>> + Send;
    fn contacts(
        &self,
    ) -> impl Future<Output = Result<Self::ContactsIter, Self::ContentsStoreError>>;
    fn contact_by_id(
        &self,
        id: &Uuid,
    ) -> impl Future<Output = Result<Option<Contact>, Self::ContentsStoreError>> + Send;
    fn clear_groups(
        &mut self,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn save_group(
        &self,
        master_key: GroupMasterKeyBytes,
        group: impl Into<Group>,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn groups(
        &self,
    ) -> impl Future<Output = Result<Self::GroupsIter, Self::ContentsStoreError>>;
    fn group(
        &self,
        master_key: GroupMasterKeyBytes,
    ) -> impl Future<Output = Result<Option<Group>, Self::ContentsStoreError>>;
    fn save_group_avatar(
        &self,
        master_key: GroupMasterKeyBytes,
        avatar: &AvatarBytes,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn group_avatar(
        &self,
        master_key: GroupMasterKeyBytes,
    ) -> impl Future<Output = Result<Option<AvatarBytes>, Self::ContentsStoreError>>;
    fn upsert_profile_key(
        &mut self,
        uuid: &Uuid,
        key: ProfileKey,
    ) -> impl Future<Output = Result<bool, Self::ContentsStoreError>> + Send;
    fn profile_key(
        &self,
        service_id: &ServiceId,
    ) -> impl Future<Output = Result<Option<ProfileKey>, Self::ContentsStoreError>> + Send;
    fn save_profile(
        &mut self,
        uuid: Uuid,
        key: ProfileKey,
        profile: Profile,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn profile(
        &self,
        uuid: Uuid,
        key: ProfileKey,
    ) -> impl Future<Output = Result<Option<Profile>, Self::ContentsStoreError>>;
    fn save_profile_avatar(
        &mut self,
        uuid: Uuid,
        key: ProfileKey,
        profile: &AvatarBytes,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>;
    fn profile_avatar(
        &self,
        uuid: Uuid,
        key: ProfileKey,
    ) -> impl Future<Output = Result<Option<AvatarBytes>, Self::ContentsStoreError>>;
    fn add_sticker_pack(
        &mut self,
        pack: &StickerPack,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>> + Send;
    fn sticker_pack(
        &self,
        id: &[u8],
    ) -> impl Future<Output = Result<Option<StickerPack>, Self::ContentsStoreError>>;
    fn remove_sticker_pack(
        &mut self,
        id: &[u8],
    ) -> impl Future<Output = Result<bool, Self::ContentsStoreError>>;
    fn sticker_packs(
        &self,
    ) -> impl Future<Output = Result<Self::StickerPacksIter, Self::ContentsStoreError>>;

    // Provided methods
    fn expire_timer(
        &self,
        thread: &Thread,
    ) -> impl Future<Output = Result<Option<(u32, u32)>, Self::ContentsStoreError>> { ... }
    fn update_expire_timer(
        &mut self,
        thread: &Thread,
        timer: u32,
        version: u32,
    ) -> impl Future<Output = Result<(), Self::ContentsStoreError>> { ... }

}

Stores messages, contacts, groups and profiles

Required Associated Types

type ContentsStoreError: StoreError

type ContactsIter: Iterator<Item = Result<Contact, Self::ContentsStoreError>>

Iterator over the contacts

type GroupsIter: Iterator<Item = Result<(GroupMasterKeyBytes, Group), Self::ContentsStoreError>>

Iterator over all stored groups

Each items is a tuple consisting of the group master key and its corresponding data.

type MessagesIter: Iterator<Item = Result<Content, Self::ContentsStoreError>>

Iterator over all stored messages

type StickerPacksIter: Iterator<Item = Result<StickerPack, Self::ContentsStoreError>>

Iterator over all stored sticker packs

Required Methods

fn clear_profiles( &mut self, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

fn clear_contents( &mut self, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

fn clear_messages( &mut self, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Clear all stored messages.

fn clear_thread( &mut self, thread: &Thread, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Clear the messages in a thread.

fn save_message( &self, thread: &Thread, message: Content, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Save a message in a Thread identified by a timestamp.

fn delete_message( &mut self, thread: &Thread, timestamp: u64, ) -> impl Future<Output = Result<bool, Self::ContentsStoreError>>

Delete a single message, identified by its received timestamp from a thread. Useful when you want to delete a message locally only.

fn message( &self, thread: &Thread, timestamp: u64, ) -> impl Future<Output = Result<Option<Content>, Self::ContentsStoreError>>

Retrieve a message from a Thread by its timestamp.

fn messages( &self, thread: &Thread, range: impl RangeBounds<u64>, ) -> impl Future<Output = Result<Self::MessagesIter, Self::ContentsStoreError>>

Retrieve all messages from a Thread within a range in time

fn clear_contacts( &mut self, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Clear all saved synchronized contact data

fn save_contact( &mut self, contacts: &Contact, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>> + Send

Save a contact

fn contacts( &self, ) -> impl Future<Output = Result<Self::ContactsIter, Self::ContentsStoreError>>

Get an iterator on all stored (synchronized) contacts

fn contact_by_id( &self, id: &Uuid, ) -> impl Future<Output = Result<Option<Contact>, Self::ContentsStoreError>> + Send

Get contact data for a single user by its Uuid.

fn clear_groups( &mut self, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Delete all cached group data

fn save_group( &self, master_key: GroupMasterKeyBytes, group: impl Into<Group>, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Save a group in the cache

fn groups( &self, ) -> impl Future<Output = Result<Self::GroupsIter, Self::ContentsStoreError>>

Get an iterator on all cached groups

fn group( &self, master_key: GroupMasterKeyBytes, ) -> impl Future<Output = Result<Option<Group>, Self::ContentsStoreError>>

Retrieve a single unencrypted group indexed by its [GroupMasterKeyBytes]

fn save_group_avatar( &self, master_key: GroupMasterKeyBytes, avatar: &AvatarBytes, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Save a group avatar in the cache

fn group_avatar( &self, master_key: GroupMasterKeyBytes, ) -> impl Future<Output = Result<Option<AvatarBytes>, Self::ContentsStoreError>>

Retrieve a group avatar from the cache.

fn upsert_profile_key( &mut self, uuid: &Uuid, key: ProfileKey, ) -> impl Future<Output = Result<bool, Self::ContentsStoreError>> + Send

Insert or update the profile key of a contact

fn profile_key( &self, service_id: &ServiceId, ) -> impl Future<Output = Result<Option<ProfileKey>, Self::ContentsStoreError>> + Send

Get the profile key for a contact

fn save_profile( &mut self, uuid: Uuid, key: ProfileKey, profile: Profile, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Save a profile by Uuid and ProfileKey.

fn profile( &self, uuid: Uuid, key: ProfileKey, ) -> impl Future<Output = Result<Option<Profile>, Self::ContentsStoreError>>

Retrieve a profile by Uuid and ProfileKey.

fn save_profile_avatar( &mut self, uuid: Uuid, key: ProfileKey, profile: &AvatarBytes, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Save a profile avatar by Uuid and ProfileKey.

fn profile_avatar( &self, uuid: Uuid, key: ProfileKey, ) -> impl Future<Output = Result<Option<AvatarBytes>, Self::ContentsStoreError>>

Retrieve a profile avatar by Uuid and ProfileKey.

fn add_sticker_pack( &mut self, pack: &StickerPack, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>> + Send

Add a sticker pack

fn sticker_pack( &self, id: &[u8], ) -> impl Future<Output = Result<Option<StickerPack>, Self::ContentsStoreError>>

Gets a cached sticker pack

fn remove_sticker_pack( &mut self, id: &[u8], ) -> impl Future<Output = Result<bool, Self::ContentsStoreError>>

Removes a sticker pack

fn sticker_packs( &self, ) -> impl Future<Output = Result<Self::StickerPacksIter, Self::ContentsStoreError>>

Get an iterator on all installed stickerpacks

Provided Methods

fn expire_timer( &self, thread: &Thread, ) -> impl Future<Output = Result<Option<(u32, u32)>, Self::ContentsStoreError>>

Get the expire timer from a Thread, which corresponds to either Contact::expire_timer or Group::disappearing_messages_timer.

fn update_expire_timer( &mut self, thread: &Thread, timer: u32, version: u32, ) -> impl Future<Output = Result<(), Self::ContentsStoreError>>

Update the expire timer from a Thread, which corresponds to either Contact::expire_timer or Group::disappearing_messages_timer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.