presage::store
Trait Store

pub trait Store:
    StateStore<StateStoreError = Self::Error>
    + ContentsStore<ContentsStoreError = Self::Error>
    + Send
    + Sync
    + Clone
    + 'static {
    type Error: StoreError;
    type AciStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone;
    type PniStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone;

    // Required methods
    fn clear(
        &mut self,
    ) -> impl Future<Output = Result<(), <Self as StateStore>::StateStoreError>> + Send;
    fn aci_protocol_store(&self) -> Self::AciStore;
    fn pni_protocol_store(&self) -> Self::PniStore;
}

The manager store trait combining all other stores into a single one

Required Associated Types

type Error: StoreError

type AciStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone

type PniStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone

Required Methods

fn clear( &mut self, ) -> impl Future<Output = Result<(), <Self as StateStore>::StateStoreError>> + Send

Clear the entire store

This can be useful when resetting an existing client.

fn aci_protocol_store(&self) -> Self::AciStore

fn pni_protocol_store(&self) -> Self::PniStore

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.