Trait presage::store::Store

pub trait Store: StateStore<StateStoreError = Self::Error> + ContentsStore<ContentsStoreError = Self::Error> + Send + Sync + Clone + 'static {
    type Error: StoreError;
    type AciStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone;
    type PniStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone;

    // Required methods
    fn clear(&mut self) -> Result<(), <Self as StateStore>::StateStoreError>;
    fn aci_protocol_store(&self) -> Self::AciStore;
    fn pni_protocol_store(&self) -> Self::PniStore;
}

The manager store trait combining all other stores into a single one

Required Associated Types

type Error: StoreError

type AciStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone

type PniStore: ProtocolStore + PreKeysStore + SenderKeyStore + SessionStoreExt + Sync + Clone

Required Methods

fn clear(&mut self) -> Result<(), <Self as StateStore>::StateStoreError>

Clear the entire store

This can be useful when resetting an existing client.

fn aci_protocol_store(&self) -> Self::AciStore

fn pni_protocol_store(&self) -> Self::PniStore

Object Safety

This trait is not object safe.