Struct presage::Manager

pub struct Manager<ConfigStore, State> { /* private fields */ }

Implementations

impl<C: ConfigStore> Manager<C, Registration>

pub async fn register(
config_store: C,
registration_options: RegistrationOptions<'_>
) -> Result<Manager<C, Confirmation>, Error>

Registers a new account with a phone number (and some options).

The returned value is a confirmation manager which you then have to use to send the confirmation code.

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    use std::str::FromStr;

    use presage::{
        prelude::{phonenumber::PhoneNumber, SignalServers},
        Manager, RegistrationOptions, SledConfigStore,
    };

    let config_store = SledConfigStore::new("/tmp/presage-example")?;

    let manager = Manager::register(
        config_store,
        RegistrationOptions {
            signal_servers: SignalServers::Production,
            phone_number: PhoneNumber::from_str("+16137827274")?,
            use_voice_call: false,
            captcha: None,
            force: false,
        },
    )
    .await?;

    Ok(())
}

impl<C: ConfigStore> Manager<C, Linking>

pub async fn link_secondary_device(
config_store: C,
signal_servers: SignalServers,
device_name: String,
provisioning_link_channel: Sender<Url>
) -> Result<Manager<C, Registered>, Error>

Links this client as a secondary device from the device used to register the account (usually a phone). The URL to present to the user will be sent in the channel given as the argument.

use presage::{prelude::SignalServers, Manager, SledConfigStore};
use futures::{channel::oneshot, future, StreamExt};

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    let config_store = SledConfigStore::new("/tmp/presage-example")?;

    let (mut tx, mut rx) = oneshot::channel();
    let (manager, err) = future::join(
        Manager::link_secondary_device(
            config_store,
            SignalServers::Production,
            "my-linked-client".into(),
            tx,
        ),
        async move {
            match rx.await {
                Ok(url) => println!("Show URL {} as QR code to user", url),
                Err(e) => println!("Error linking device: {}", e)
            }
        },
    )
    .await;

    Ok(())
}

impl<C: ConfigStore> Manager<C, Confirmation>

pub async fn confirm_verification_code(
self,
confirm_code: u32
) -> Result<Manager<C, Registered>, Error>

Confirm a newly registered account using the code you received by SMS or phone call.

Returns a registered manager that you can use to send and receive messages.

impl<C: ConfigStore> Manager<C, Registered>

pub fn load_registered(config_store: C) -> Result<Self, Error>

Loads a previously registered account from the implemented ConfigStore.

Returns a instance of Manager you can use to send & receive messages.

pub async fn request_contacts_sync(&self) -> Result<(), Error>

Request that the primary device to encrypt & send all of its contacts as a message to ourselves which can be then received, decrypted and stored in the message receiving loop.

Note: If successful, the contacts are not yet received and stored, but will only be processed when they’re received using the MessageReceiver.

pub fn state(&self) -> &Registered

Returns a handle on the registered state

pub fn uuid(&self) -> Uuid

Get the profile UUID

pub async fn whoami(&self) -> Result<WhoAmIResponse, Error>

Fetches basic information on the registered device.

pub async fn retrieve_profile(&self) -> Result<Profile, Error>

Fetches the profile (name, about, status emoji) of the registered user.

pub async fn retrieve_profile_by_uuid(
&self,
uuid: Uuid,
profile_key: [u8; 32]
) -> Result<Profile, Error>

Fetches the profile of the provided user by UUID and profile key.

pub fn get_contacts(&self) -> Result<impl Iterator<Item = Contact>, Error>

Returns an iterator on contacts stored in the ConfigStore.

Note: after requesting contacts sync, you need to start the receiving message loop for contacts to be processed

pub fn get_contact_by_id(&self, id: Uuid) -> Result<Option<Contact>, Error>

pub async fn receive_messages(
&self
) -> Result<impl Stream<Item = Content>, Error>

Starts receiving messages.

Returns a [Stream] of messages to consume.

Note: messages that aren’t consumed will be gone, as internally we acknknowledge all messages upon receival.

pub async fn send_message(
&self,
recipient_addr: impl Into<ServiceAddress>,
message: impl Into<ContentBody>,
timestamp: u64
) -> Result<(), Error>

Sends a messages to the provided ServiceAddress. The timestamp should be set to now and is used by Signal mobile apps to order messages later, and apply reactions.

pub async fn upload_attachments(
&self,
attachments: Vec<(AttachmentSpec, Vec<u8>)>
) -> Result<Vec<Result<AttachmentPointer, AttachmentUploadError>>, Error>

Uploads attachments prior to linking them in a message.

pub async fn send_message_to_group(
&self,
recipients: impl IntoIterator<Item = ServiceAddress>,
message: DataMessage,
timestamp: u64
) -> Result<(), Error>

Sends one message in a group (v2).

pub async fn clear_sessions(
&self,
recipient: &ServiceAddress
) -> Result<(), Error>

Clears all sessions established wiht recipient.

pub async fn get_group_v2(
&self,
group_master_key: GroupMasterKey
) -> Result<Group, Error>

pub fn decrypt_group_context(
&self,
group_context: GroupContextV2
) -> Result<Option<GroupChanges>, Error>

Decrypts a blob of GroupContextV2 and deserializes it in a higher level GroupChanges struct.

pub async fn get_attachment(
&self,
attachment_pointer: &AttachmentPointer
) -> Result<Vec<u8>, Error>

Downloads and decrypts a single attachment.