Trait presage::MessageStore

pub trait MessageStore {
    type MessagesIter: Iterator<Item = Result<Content, Error>>;

    fn save_message(
        &mut self,
        thread: &Thread,
        message: Content
    ) -> Result<(), Error>;
    fn delete_message(
        &mut self,
        thread: &Thread,
        timestamp: u64
    ) -> Result<bool, Error>;
    fn message(
        &self,
        thread: &Thread,
        timestamp: u64
    ) -> Result<Option<Content>, Error>;
    fn messages(
        &self,
        thread: &Thread,
        from: Option<u64>
    ) -> Result<Self::MessagesIter, Error>;
}

A MessageStore can store messages in the form Content and retrieve messages either by [MessageIdentity], by Thread or completely.

Required Associated Types

type MessagesIter: Iterator<Item = Result<Content, Error>>

Required Methods

fn save_message(&mut self, thread: &Thread, message: Content) -> Result<(), Error>

Save a message in a Thread identified by a timestamp. TODO: deriving the thread happens from the content, so we can also ditch the first parameter

fn delete_message(
&mut self,
thread: &Thread,
timestamp: u64
) -> Result<bool, Error>

Delete a single message, identified by its received timestamp from a thread.

fn message(
&self,
thread: &Thread,
timestamp: u64
) -> Result<Option<Content>, Error>

Retrieve a message from a Thread by its timestamp.

fn messages(
&self,
thread: &Thread,
from: Option<u64>
) -> Result<Self::MessagesIter, Error>

Retrieve a message from a Thread.