Trait presage::Store

pub trait Store: ProtocolStore + SenderKeyStore + SessionStoreExt + Sync + Clone {
    type Error: StoreError;
    type ContactsIter: Iterator<Item = Result<Contact, Self::Error>>;
    type GroupsIter: Iterator<Item = Result<(GroupMasterKeyBytes, Group), Self::Error>>;
    type MessagesIter: Iterator<Item = Result<Content, Self::Error>>;

    // Required methods
    fn load_state(&self) -> Result<Option<Registered>, Self::Error>;
    fn save_state(&mut self, state: &Registered) -> Result<(), Self::Error>;
    fn is_registered(&self) -> bool;
    fn clear_registration(&mut self) -> Result<(), Self::Error>;
    fn clear(&mut self) -> Result<(), Self::Error>;
    fn pre_keys_offset_id(&self) -> Result<u32, Self::Error>;
    fn set_pre_keys_offset_id(&mut self, id: u32) -> Result<(), Self::Error>;
    fn next_signed_pre_key_id(&self) -> Result<u32, Self::Error>;
    fn next_pq_pre_key_id(&self) -> Result<u32, Self::Error>;
    fn set_next_signed_pre_key_id(&mut self, id: u32) -> Result<(), Self::Error>;
    fn set_next_pq_pre_key_id(&mut self, id: u32) -> Result<(), Self::Error>;
    fn clear_messages(&mut self) -> Result<(), Self::Error>;
    fn clear_thread(&mut self, thread: &Thread) -> Result<(), Self::Error>;
    fn save_message(
        &mut self,
        thread: &Thread,
        message: Content
    ) -> Result<(), Self::Error>;
    fn delete_message(
        &mut self,
        thread: &Thread,
        timestamp: u64
    ) -> Result<bool, Self::Error>;
    fn message(
        &self,
        thread: &Thread,
        timestamp: u64
    ) -> Result<Option<Content>, Self::Error>;
    fn messages(
        &self,
        thread: &Thread,
        range: impl RangeBounds<u64>
    ) -> Result<Self::MessagesIter, Self::Error>;
    fn clear_contacts(&mut self) -> Result<(), Self::Error>;
    fn save_contacts(
        &mut self,
        contacts: impl Iterator<Item = Contact>
    ) -> Result<(), Self::Error>;
    fn contacts(&self) -> Result<Self::ContactsIter, Self::Error>;
    fn contact_by_id(&self, id: Uuid) -> Result<Option<Contact>, Self::Error>;
    fn clear_groups(&mut self) -> Result<(), Self::Error>;
    fn save_group(
        &self,
        master_key: GroupMasterKeyBytes,
        group: &Group
    ) -> Result<(), Self::Error>;
    fn groups(&self) -> Result<Self::GroupsIter, Self::Error>;
    fn group(
        &self,
        master_key: GroupMasterKeyBytes
    ) -> Result<Option<Group>, Self::Error>;
    fn upsert_profile_key(
        &mut self,
        uuid: &Uuid,
        key: ProfileKey
    ) -> Result<bool, Self::Error>;
    fn profile_key(
        &self,
        uuid: &Uuid
    ) -> Result<Option<ProfileKey>, Self::Error>;
    fn save_profile(
        &mut self,
        uuid: Uuid,
        key: ProfileKey,
        profile: Profile
    ) -> Result<(), Self::Error>;
    fn profile(
        &self,
        uuid: Uuid,
        key: ProfileKey
    ) -> Result<Option<Profile>, Self::Error>;

    // Provided method
    fn expire_timer(&self, thread: &Thread) -> Result<Option<u32>, Self::Error> { ... }

}

Required Associated Types

type Error: StoreError

type ContactsIter: Iterator<Item = Result<Contact, Self::Error>>

type GroupsIter: Iterator<Item = Result<(GroupMasterKeyBytes, Group), Self::Error>>

type MessagesIter: Iterator<Item = Result<Content, Self::Error>>

Required Methods

fn load_state(&self) -> Result<Option<Registered>, Self::Error>

State Load registered (or linked) state

fn save_state(&mut self, state: &Registered) -> Result<(), Self::Error>

Save registered (or linked) state

fn is_registered(&self) -> bool

Returns whether this store contains registration data or not

fn clear_registration(&mut self) -> Result<(), Self::Error>

Clear registration data (including keys), but keep received messages, groups and contacts.

fn clear(&mut self) -> Result<(), Self::Error>

Clear the entire store: this can be useful when resetting an existing client.

fn pre_keys_offset_id(&self) -> Result<u32, Self::Error>

Pre-keys

fn set_pre_keys_offset_id(&mut self, id: u32) -> Result<(), Self::Error>

fn next_signed_pre_key_id(&self) -> Result<u32, Self::Error>

fn next_pq_pre_key_id(&self) -> Result<u32, Self::Error>

fn set_next_signed_pre_key_id(&mut self, id: u32) -> Result<(), Self::Error>

fn set_next_pq_pre_key_id(&mut self, id: u32) -> Result<(), Self::Error>

fn clear_messages(&mut self) -> Result<(), Self::Error>

Messages

fn clear_thread(&mut self, thread: &Thread) -> Result<(), Self::Error>

fn save_message( &mut self, thread: &Thread, message: Content ) -> Result<(), Self::Error>

Save a message in a Thread identified by a timestamp.

fn delete_message( &mut self, thread: &Thread, timestamp: u64 ) -> Result<bool, Self::Error>

👎Deprecated: message deletion is now handled internally

Delete a single message, identified by its received timestamp from a thread.

fn message( &self, thread: &Thread, timestamp: u64 ) -> Result<Option<Content>, Self::Error>

Retrieve a message from a Thread by its timestamp.

fn messages( &self, thread: &Thread, range: impl RangeBounds<u64> ) -> Result<Self::MessagesIter, Self::Error>

Retrieve all messages from a Thread within a range in time

fn clear_contacts(&mut self) -> Result<(), Self::Error>

Contacts Clear all saved synchronized contact data

fn save_contacts( &mut self, contacts: impl Iterator<Item = Contact> ) -> Result<(), Self::Error>

Replace all contact data

fn contacts(&self) -> Result<Self::ContactsIter, Self::Error>

Get an iterator on all stored (synchronized) contacts

fn contact_by_id(&self, id: Uuid) -> Result<Option<Contact>, Self::Error>

Get contact data for a single user by its Uuid.

fn clear_groups(&mut self) -> Result<(), Self::Error>

Delete all cached group data

fn save_group( &self, master_key: GroupMasterKeyBytes, group: &Group ) -> Result<(), Self::Error>

Save a group in the cache

fn groups(&self) -> Result<Self::GroupsIter, Self::Error>

Get an iterator on all cached groups

fn group( &self, master_key: GroupMasterKeyBytes ) -> Result<Option<Group>, Self::Error>

Retrieve a single unencrypted group indexed by its [GroupMasterKeyBytes]

fn upsert_profile_key( &mut self, uuid: &Uuid, key: ProfileKey ) -> Result<bool, Self::Error>

Profiles Insert or update the profile key of a contact

fn profile_key(&self, uuid: &Uuid) -> Result<Option<ProfileKey>, Self::Error>

Get the profile key for a contact

fn save_profile( &mut self, uuid: Uuid, key: ProfileKey, profile: Profile ) -> Result<(), Self::Error>

Save a profile by Uuid and ProfileKey.

fn profile( &self, uuid: Uuid, key: ProfileKey ) -> Result<Option<Profile>, Self::Error>

Retrieve a profile by Uuid and ProfileKey.

Provided Methods

fn expire_timer(&self, thread: &Thread) -> Result<Option<u32>, Self::Error>

Get the expire timer from a Thread, which corresponds to either Contact::expire_timer or Group::disappearing_messages_timer.